Moves Categories listing classes to list package
