--- WordPress/src/main/java/org/wordpress/android/ui/prefs/categories/Utils.kt
@@ -1,5 +1,0 @@
-package org.wordpress.android.ui.prefs.categories
-
-sealed class CategoryDetailNavigation
-object CreateCategory : CategoryDetailNavigation()
-data class EditCategory(val categoryId: Long) : CategoryDetailNavigation()
