+++ WordPress/src/main/java/org/wordpress/android/ui/prefs/categories/list/SiteSettingsCategoriesViewHolder.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui.prefs.categories
+package org.wordpress.android.ui.prefs.categories.list
 
 import android.os.Bundle
 import androidx.core.view.ViewCompat
