+++ WordPress/src/test/java/org/wordpress/android/ui/prefs/categories/CategoriesListViewModelTest.kt
@@ -18,11 +18,12 @@ import org.wordpress.android.fluxc.store.TaxonomyStore.OnTaxonomyChanged
 import org.wordpress.android.fluxc.store.TaxonomyStore.TaxonomyError
 import org.wordpress.android.fluxc.store.TaxonomyStore.TaxonomyErrorType.GENERIC_ERROR
 import org.wordpress.android.ui.posts.GetCategoriesUseCase
-import org.wordpress.android.ui.prefs.categories.CategoriesListViewModel.UiState
-import org.wordpress.android.ui.prefs.categories.CategoriesListViewModel.UiState.Content
-import org.wordpress.android.ui.prefs.categories.CategoriesListViewModel.UiState.Error.GenericError
-import org.wordpress.android.ui.prefs.categories.CategoriesListViewModel.UiState.Error.NoConnection
-import org.wordpress.android.ui.prefs.categories.CategoriesListViewModel.UiState.Loading
+import org.wordpress.android.ui.prefs.categories.list.CategoriesListViewModel
+import org.wordpress.android.ui.prefs.categories.list.UiState
+import org.wordpress.android.ui.prefs.categories.list.UiState.Content
+import org.wordpress.android.ui.prefs.categories.list.UiState.Error.GenericError
+import org.wordpress.android.ui.prefs.categories.list.UiState.Error.NoConnection
+import org.wordpress.android.ui.prefs.categories.list.UiState.Loading
 import org.wordpress.android.util.NetworkUtilsWrapper
 
 @InternalCoroutinesApi
